Removed default panel header.  I think keeping...
the header is useful for now.  We are using it in both the
events panel and the list panel, which makes sense because the
"content_container" of those panels is populated with content,
while the header is used for something else (i.e. an input form).
Also the per/panel loading notification is useful currently.